third_party/blink/web_tests/platform/generic/html/grouping_content/figcaption-element-expected.txt

Various tests for the figcaption element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

<figcaption> default styling:
PASS getStyleValue("figcaption0","display") is "block"
<figcaption> closes <p>:
PASS figcaption1.parentNode.nodeName == "p" is false
<p> does not close <figcaption>:
PASS p1.parentNode.nodeName is "FIGCAPTION"
<figcaption> can be nested inside <figcaption> or <footer>:
PASS figcaption3.parentNode.id is "figcaption2"
PASS figcaption5.parentNode.id is "footer1"
Residual style:
PASS getStyleValue("figcaption4","font-weight") is "700"
PASS getStyleValue("span1","font-weight") is "700"
PASS successfullyParsed is true

TEST COMPLETE
